server/requirements.txt
Django==2.0.6
dj-database-url==0.4.1
gunicorn==19.8.1
ipdb==0.9.3
ipython==4.2.0
psycopg2-binary==2.7.4
djangorestframework==3.8.2
rest-framework-generic-relations==1.1.0
django-debug-toolbar==1.9.1
django-rest-swagger==2.2.0
django-cors-headers==2.4.0
celery==4.2.0
redis==2.10.6
django-redis==4.9.0
requests==2.9.1
